add second bg, change caracter prefabs, change upgrade desk prefabs, start work with achievements

+++ iBelieve/Assets/Scripts/UI/Cheats.cs
@@ -10,7 +10,7 @@ public class Cheats : MonoBehaviour
     [SerializeField] private GameObject field;
     [SerializeField] private Button useButton;
     [SerializeField] private Button closeButton;
-    private int curValue;
+    private float curValue;
     private Coins coins;
     void Start()
     {
@@ -34,9 +34,9 @@ public class Cheats : MonoBehaviour
 
     private void UseCheats()
     {
-        if (int.TryParse(field.GetComponent<TMP_InputField>().text, out curValue))
+        if (float.TryParse(field.GetComponent<TMP_InputField>().text, out curValue))
         {
-            curValue = int.Parse(field.GetComponent<TMP_InputField>().text);
+            curValue = float.Parse(field.GetComponent<TMP_InputField>().text);
             coins.AddCoins(curValue);
             field.GetComponent<TMP_InputField>().text = "";
             cheatsPanel.SetActive(false);
